[CK_TILE] Add bf16 support to GEMM Tile-Engine -> Dispatcher bridge
Closes the regular-path dtype-coverage gap (fp16-only) identified when
measuring the GEMM bridge against the FMHA reference (PR #5260). The C ABI
is already dtype-agnostic (sizes buffers from sizeof(ADataType)) and the
codegen already emits bf16 kernels; the only fp16 hard-coding lived in the
Python host buffers.

- gemm_utils.py: GpuGemmRunner.run() is now dtype-aware. It detects the
  kernel's real dtype from the compiled kernel name and encodes the host
  buffers to match. bf16 has no native numpy dtype, so it is carried as a
  uint16 bit pattern (sizeof(bf16_t) == 2): _fp32_to_bf16_u16 encodes
  fp32 -> bf16 with round-to-nearest-even, _bf16_u16_to_fp32 decodes the
  output back to fp32. fp16 path is unchanged.
- run_one_gemm_kernel.py: worker now generates fp32 source data so the
  runner owns all dtype encoding (no double rounding).

Stacked on muozturk/dispatcher-gemm-registry-key (PR #8187): bf16 needs the
trait-derived KernelKey so the registry reports the correct dtype.

Validated on gfx942 (MI300X): 3 bf16 kernels x 2 problems = 6/6 OK,
46-219 TFLOPS; numeric check vs CPU reference max_rel_err 7.7e-3 (within
bf16 tolerance).

+++ projects/composablekernel/tile_engine/ops/gemm/run_one_gemm_kernel.py
@@ -37,8 +37,10 @@ def _run_one(idx, so_path, prob_dict, kernel_name):
         problem = GemmProblem.from_dict(prob_dict)
 
         np.random.seed(42)
-        A = (np.random.randn(problem.M, problem.K) * 0.1).astype(np.float16)
-        B = (np.random.randn(problem.K, problem.N) * 0.1).astype(np.float16)
+        # Generate fp32 source; the runner encodes to the kernel's real dtype
+        # (fp16 or bf16) based on the compiled kernel name.
+        A = (np.random.randn(problem.M, problem.K) * 0.1).astype(np.float32)
+        B = (np.random.randn(problem.K, problem.N) * 0.1).astype(np.float32)
 
         # CRITICAL: load the library ONLY inside this subprocess.
         runner = GpuGemmRunner(lib_path=so_path)
